+++ modules/disk-io.c
@@ -310,7 +310,7 @@ disk_io_new(uint16_t interval, struct particle *label)
 static struct module *
 from_conf(const struct yml_node *node, struct conf_inherit inherited)
 {
-    const struct yml_node *interval = yml_get_value(node, "interval");
+    const struct yml_node *interval = yml_get_value(node, "poll-interval");
     const struct yml_node *c = yml_get_value(node, "content");
 
     return disk_io_new(
@@ -326,7 +326,7 @@ conf_verify_interval(keychain_t *chain, const struct yml_node *node)
 
     if (yml_value_as_int(node) < SMALLEST_INTERVAL) {
         LOG_ERR(
-            "%s: interval value cannot be less than %d ms",
+            "%s: poll-interval value cannot be less than %d ms",
             conf_err_prefix(chain, node), SMALLEST_INTERVAL);
         return false;
     }
@@ -338,7 +338,7 @@ static bool
 verify_conf(keychain_t *chain, const struct yml_node *node)
 {
     static const struct attr_info attrs[] = {
-        {"interval", false, &conf_verify_interval},
+        {"poll-interval", false, &conf_verify_interval},
         MODULE_COMMON_ATTRS,
     };